requirements.txt
pycryptodome
bitarray